MDEV-9382: After updating mariadb server apt-configure fails
When acting as a Galera receiver node, server startup may take
more than 30 secs (the current default) as it has to wait for
SST/IST operation to complete besides spending some time doing
wsrep recovery.

In specific case of this MDEV, the server startup timedout as
it took ~48 secs for the node to fully start.

A viable workaround would be to adjust the MYSQLD_STARTUP_TIMEOUT
env to a higher value, but raising the default to 60 secs should
fix scenarios like these.